+++ ctapipe/flow/flow.py
@@ -15,7 +15,7 @@ from ctapipe.flow.multiprocessus.router_queue_zmq import RouterQueue
 from ctapipe.flow.sequential.producer_sequential import ProducerSequential
 from ctapipe.flow.sequential.stager_sequential import StagerSequential
 from ctapipe.flow.sequential.consumer_sequential import ConsumerSequential
-from ctapipe.flow.gui.graphwidget import StagerRep
+from ctapipe.flow.stager_rep import StagerRep
 from ctapipe.utils import dynamic_class_from_module
 from ctapipe.utils.dynamic_class import DynamicClassError
 from ctapipe.core import Tool
